refactor: extract autosuggest and batch input components
